requirements.txt
discord==2.2.2
PyNaCl==1.5.0
python-dotenv>=1.0.0